Fotos y datos de las leyendes en firebase

+++ AdivinaDondeEstoyProyecto/app/src/main/java/Modelo/Leyend.kt
@@ -1,3 +1,3 @@
 package Modelo
 
-class Leyend (var nombre: String, var cordenadaX: Int, var cordenadaY: Int, var historia: String, var acertado: Boolean)
+class Leyend (var nombre: String, var cordenadaX: Float, var cordenadaY: Float, var historia: String, var acertado: Boolean)
